Swapping it all out for the "latest" "stable" version
Update .gitignore

Update .gitignore

Deleted a whole bunch'a shit

Update LICENSE.md

Create pull-files-from-azure-storage.yml

Update push-files-to-azure-storage.yml

Update push-files-to-azure-storage.yml

Update and rename .github/workflows/.old/push_files_to_azure_storage.yml to .github/workflows/push-files-to-azure-storage.yml

Add or update the App Service deployment workflow configuration from Azure Portal.

Update docker-compose.yml

Lemme try one more thing...

Add or update the App Service deployment workflow configuration from Azure Portal.

Add or update the App Service deployment workflow configuration from Azure Portal.

Create an auto-deploy file

Update pull-the-repo-and-deploy-directly-within-ACR.yml

Update pull-the-repo-and-deploy-directly-within-ACR.yml

Think I might  finally've solved it!

Update move-and-populate-final-production-ready-files.yml

Update move-and-populate-final-production-ready-files.yml

Update pull-the-repo-and-deploy-directly-within-ACR.yml

Update pull-the-repo-and-deploy-directly-within-ACR.yml

Update pull-the-repo-and-deploy-directly-within-ACR.yml

Update pull-the-repo-and-deploy-directly-within-ACR.yml

Update pull-the-repo-and-deploy-directly-within-ACR.yml

Update pull-the-repo-and-deploy-directly-within-ACR.yml

Update build-and-deploy-the-whole-shebang.yml

Update build-and-deploy-the-whole-shebang.yml

Rename build-and-deploy-from-docker-image.yml to build-and-deploy-the-whole-shebang.yml

Update build-and-deploy-from-docker-image.yml

muchos cambios

Update deploy-the-image-to-the-web-app.yml

Delete frontmatter.json

Delete .frontmatter/content directory

Update .gitignore

renaming

Update move-and-populate-final-production-ready-files.yml

Update move-and-populate-final-production-ready-files.yml

Update move-and-populate-final-production-ready-files.yml

Update move-and-populate-final-production-ready-files.yml

Update move-and-populate-final-production-ready-files.yml

Update build-and-deploy-from-docker-image.yml

Create move-and-populate-final-production-ready-files.yml

Update build-and-deploy-from-docker-image.yml

Create deploy-the-image-to-the-web-app.yml

Update build-and-deploy-from-docker-image.yml

Update build-and-deploy-from-docker-image.yml

Update build-and-deploy-from-docker-image.yml

Update and rename deploy-the-image-to-the-web-app.yml to pull-the-repo-and-deploy-directly-within-ACR.yml

Update and rename .github/workflows/build-and-push-docker-image-to-azure.yml to .github/pull-the-repo-and-deploy-directly-within-ACR.yml

Update build-and-deploy-from-docker-image.yml

Update build-and-deploy-from-docker-image.yml

fuck i hope this workd

Update build-and-push-docker-image-to-azure.yml

dunno if this'll wokr but we can try

separate dockerfules into dev & prod versins

muchos cambiod

Update build-and-push-docker-image-to-azure.yml

Update build-and-push-docker-image-to-azure.yml

Update .env

.env

Update docker-compose.yml

Update build-and-deploy-from-docker-image.yml

Update deploy-the-image-to-the-web-app.yml

Update deploy-the-image-to-the-web-app.yml

Update deploy-the-image-to-the-web-app.yml

Update build-and-push-docker-image-to-azure.yml

Update deploy-the-image-to-the-web-app.yml

Update deploy-the-image-to-the-web-app.yml

Update deploy-the-image-to-the-web-app.yml

Update deploy-the-image-to-the-web-app.yml

Update build-and-push-docker-image-to-azure.yml

Update build-and-push-docker-image-to-azure.yml

Update deploy-the-image-to-the-web-app.yml

Update deploy-the-image-to-the-web-app.yml

Update deploy-the-image-to-the-web-app.yml

Update deploy-the-image-to-the-web-app.yml

Update deploy-the-image-to-the-web-app.yml

Update deploy-the-image-to-the-web-app.yml

Think I might've finally solved it!

moving older (not needed/used) workflow files to a .old folder

fixing the missing folder problems

Update deploy-the-image-to-the-web-app.yml

parameterizing the APACHE_DOCUMENT_ROOT directory

Update build-and-push-docker-image-to-azure.yml

Moving these to the root of the repo

spelling "docker" right....

adding docker-compose to the docker image

Update deploy-the-image-to-the-web-app.yml

Update build-and-deploy-from-docker-image.yml

Update send-webhook-to-server.yml

Update deploy-the-image-to-the-web-app.yml

Update build-and-push-docker-image-to-azure.yml

add WP CLI

Update docker-compose.yml

Update Dockerfile

Update deploy-the-image-to-the-web-app.yml

Update Dockerfile

Update build-and-push-docker-image-to-azure.yml

Think I might've finally solved it!

Update deploy-the-image-to-the-web-app.yml

Update send-webhook-to-server.yml

remove call to startup.sh from docker-compose

Update startup.sh

un-sudo'ing everything

UPDATE DOCKERFILE TO INSTALLL GH CLI

getting rid of nginx and certbot

sudo'ing all this stuff again

Changing dockerfile & adding startup.sh

removing obsolete .git files

Update push_files_to_azure_storage.yml

changing dockerfile back to how it was

changing Dockerfile

changing Dockerfile

changing Dockerfile

Move home html folder

Update send-webhook-to-server.yml

Update send-webhook-to-server.yml

Create send-webhook-to-server.yml

Update build-and-deploy-from-docker-image.yml

Update build-and-push-docker-image-to-azure.yml

Update push_files_to_azure_storage.yml

Update build-and-deploy-from-docker-image.yml

Update build-and-deploy-from-docker-image.yml

Update build-and-deploy-from-docker-image.yml

Update build-and-deploy-from-docker-image.yml

Update build-and-deploy-from-docker-image.yml

Update and rename build_and_push_docker_image_to_azure.yml to build-and-push-docker-image-to-azure.yml

Update deploy-the-image-to-the-web-app.yml

Update deploy-the-image-to-the-web-app.yml

Create push_files_to_azure_storage.yml

Create deploy-the-image-to-the-web-app.yml

Rename main.yml to build_and_push_docker_image_to_azure.yml

Create main.yml

Update build-and-deploy-from-docker-image.yml

Update build-and-deploy-from-docker-image.yml

Update build-and-deploy-from-docker-image.yml

Update build-and-deploy-from-docker-image.yml

Update build-and-deploy-from-docker-image.yml

Update build-and-deploy-from-docker-image.yml

Update build-and-deploy-from-docker-image.yml

Update build-and-deploy-from-docker-image.yml

Update build-and-deploy-from-docker-image.yml

Update build-and-deploy-from-docker-image.yml

Update build-and-deploy-from-docker-image.yml

Update build-and-deploy-from-docker-image.yml

Update build-and-deploy-from-docker-image.yml

Update build-and-deploy-from-docker-image.yml

Update build-and-deploy-from-docker-image.yml

Update build-and-deploy-from-docker-image.yml

Update build-and-deploy-from-docker-image.yml

Update build-and-deploy-from-docker-image.yml

Update build-and-deploy-from-docker-image.yml

Create build-and-deploy-from-docker-image.yml

Dockerfile & docker-compose.yml

turn off debugging

re-adding ACF Pro

swapping everything out for the "latest" "stable" version